Fixes display of users in workload table
The user will be displayed in a group only if it is her main group. This
feature was lost due to some minor refactoring.

The order of users displayed in the workload table is also fixed since
the dummy user was shifted to the tables end but needs to be on the
first position. This feature was lost due to a renaming of the
GroupUserDummy class infuencing the alphabetical order against the User
class.