Changes group member vacation strategy
Will consider the workload of the group user dummy when all members are
in vacation. The group workload capacity will be zero when no group
member is able to work leading to a red day field background when the
group user dummy holds open tickets with unfinished scheduled estimated
hours.