Changes unscheduled issue definition
An issue will be unscheduled when it has no due date.

This changes the prior definition of having either no due date or
start date. Since there is already a check for an issue of beeing
overdue what again implies an existing due date we will make sure that
there is no double counting of issues beeing overdue and unscheduled
when only the start date is missing.

This will also improve the algorithms efficiency.