Fixes plugin setting for general_workday_sunday
The plugin setting was accidentally overridden during a change in
[https://github.com/xmera-circle/redmine_workload/commit/7cb6773cba8b6c6223358d52e37921944c736f8f](7cb6773).

Changing the settings back again, however, raises some other errors
with RedmineWorkload::WlDateTools.holiday?. It turns out that beginless
and endless ranges does not work with Ruby < 3.

These ranges were introduced due to problems with postgres when using
'start' and 'end' as String in an SQL query.

This again leads to the
requirement for postres users to use Ruby > 3 with this plugin!